+++ CHANGELOG.md
@@ -6,10 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
 * Create Recipe Nursing Series
 
-=======
+* Remove numbering from figures without id
+
 ## [v1.25.0] - 2023-02-24
 
 * Add `top_title` behavior to `BakeUnnumberedTables`