+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Change `written-exercises` into `written-questions` in `bca`
 * Bake `full-width` tables
 * Bake `injected-exercises` in `bca`
 * Create Recipe Nursing Series